+++ guillotina_stripe/models.py
@@ -24,10 +24,11 @@ class BillingDetails(BaseModel):
 
 class Subscription(BaseModel):
     id: str
-    default_payment_method: str
+    default_payment_method: Optional[str]
     current_period_end: int
     current_period_start: int
     status: str
+    cancel_at_period_end: bool
     items: Any
     latest_invoice: Any