+++ guillotina_stripe/tests/fixtures.py
@@ -4,12 +4,14 @@ from guillotina import testing
 import pytest
 import os
 
+
 def base_settings_configurator(settings):
     if 'applications' in settings:
         settings['applications'].append('guillotina_stripe')
         settings['applications'].append('guillotina_stripe.test_package')
     else:
-        settings['applications'] = ['guillotina_stripe', 'guillotina_stripe.test_package']
+        settings['applications'] = [
+            'guillotina_stripe', 'guillotina_stripe.test_package']
     settings['load_utilities']['stripe'] = {
         "provides": "guillotina_stripe.interfaces.IStripePayUtility",
         "factory": "guillotina_stripe.utility.StripePayUtility",
@@ -37,6 +39,7 @@ settings['stripe'] = {
     'version': 'latest',
 }
 
+
 class STRIPE(BaseImage):
     name = 'stripe'
     port = 12112
@@ -59,11 +62,9 @@ class STRIPE(BaseImage):
 
 stripe_image = STRIPE()
 
+
 @pytest.fixture(scope='session')
 def stripe():
     result = stripe_image.run()
     yield result
     stripe_image.stop()
-
-
-